+++ app/Http/Controllers/Backend/ZoneController.php
@@ -132,11 +132,25 @@ class ZoneController extends Controller
 
         $id         = Input::get('selected_checkboxes');
         $new_string = explode(',', $id);
+        $delete_flag = true;
         foreach($new_string as $id){
-            $this->repo->delete($id);
+            $check = $this->repo->checkToDelete($id);
+            if(isset($check) && count($check)>0){
+                alert()->warning('There are car_type_setups with this zone_id = '.$id)->persistent('OK');
+                $delete_flag = false;
+            }
+            else{
+                $this->repo->delete($id);
+            }
         }
-        return redirect()->action('Backend\ZoneController@index')
-            ->withMessage(FormatGenerator::message('Success', 'Zones deleted ...'));
-    }
 
+        if($delete_flag){
+            return redirect()->action('Backend\ZoneController@index')
+                ->withMessage(FormatGenerator::message('Success', 'Zones deleted ...'));
+        }
+        else{
+            return redirect()->action('Backend\ZoneController@index')
+                ->withMessage(FormatGenerator::message('Fail', 'Zone did not delete ...'));
+        }
+    }
 }
